Refactor: completed separate file for S3 operations functions
